requirements.txt
spotipy==2.23.0
groq==0.4.2
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
